Rename files to avoid naming conflict with existing state.ts file
